requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile '.\requirements.in'
#
appdirs==1.4.4
    # via black
black==20.8b1
    # via -r .\requirements.in
click==7.1.2
    # via
    #   black
    #   pip-tools
    #   uvicorn
colorama==0.4.4
    # via uvicorn
dnspython==2.0.0
    # via
    #   -r .\requirements.in
    #   email-validator
email-validator==1.1.1
    # via pydantic
fastapi-users[mongodb]==3.1.1
    # via -r requirements.in
fastapi==0.61.1
    # via -r .\requirements.in
flake8==3.8.4
    # via -r .\requirements.in
h11==0.11.0
    # via uvicorn
httptools==0.1.1
    # via uvicorn
idna==2.10
    # via email-validator
makefun==1.9.5
    # via fastapi-users
mccabe==0.6.1
    # via flake8
motor==2.3.0
    # via -r .\requirements.in
mypy-extensions==0.4.3
    # via black
pathspec==0.8.0
    # via black
pip-tools==5.3.1
    # via -r .\requirements.in
pycodestyle==2.6.0
    # via flake8
pycparser==2.20
    # via cffi
pydantic[email]==1.7
    # via
    #   -r .\requirements.in
    #   fastapi
pyflakes==2.2.0
    # via flake8
pyjwt==1.7.1
    # via fastapi-users
pymongo==3.11.0
    # via motor
python-dotenv==0.14.0
    # via uvicorn
python-multipart==0.0.5
    # via
    #   -r requirements.in
    #   fastapi-users
pyyaml==5.3.1
    # via uvicorn
regex==2020.10.23
    # via black
six==1.15.0
    # via pip-tools
starlette==0.13.6
    # via fastapi
toml==0.10.1
    # via black
typed-ast==1.4.1
    # via black
typing-extensions==3.7.4.3
    # via black
ujson==4.0.1
    # via -r requirements.in
uvicorn[standard]==0.12.2
    # via -r .\requirements.in
#uvloop==0.14.0
    # via uvicorn
watchgod==0.6
    # via uvicorn
websockets==8.1
    # via uvicorn
python-jose[cryptography]

passlib[bcrypt]==1.7.4

# The following packages are considered to be unsafe in a requirements file:
# pip
#fastapi-mail==0.4.1

requests==2.26.0

aiofiles==0.8.0

#Package for Unit test case
pytest==6.2.3
trio==0.19.0